Bugfix: Trash path option not correctly handled
The trash path was not correctly handled as it was not used in deleting the real files. This has been fixed now.
The dryrun information now prints '[Remove]' in front of files it would have removed but which are not there (for debugging).
The days argument is now checked for validity (as integer) by OptParse.
All uses of a '/' have been replaced with os.path.join instead.