Added the --max-free option
The --max-free option is a new option in line with two other planned options,
called --free and --min-free. --free will try to remove the given number of
megabytes, while --min-free will automatically set --free to the difference
between the requested min-free and the actual min-free (delete files as
we run out of disk space).