Added -D, --delete-first option
Delete first allows you to run a regex that will push
files to the top of the trash queue, removing those files first.

For example, use:
 --delete-first '.*.avi' to delete *.avi files first
NOTE: the argument is a REGEX not a glob pattern.

Current implementation will delete the last match first
but this will be changed before the 0.1.1 release.