Quickfix: handle read-only files in trash
The trash can contain read-only files, for which
the permission has to be reset before it is removed.

This quickfix will add write permission and try again.
It may become the real fix, but I would like to still
restore the permissions if it does not help with deleting
them (for the rare case that between the chmod and the
unlink, something about the file changes which gives a
new permission error. Then the flags should be reset)